+++ packages/angular-workspace/spright-angular/chat/message/spright-chat-message.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { SprightChatMessageDirective } from './spright-chat-message.directive';
+
+import '@ni/spright-components/dist/esm/chat/message';
+
+@NgModule({
+    declarations: [
+        SprightChatMessageDirective
+    ],
+    imports: [CommonModule],
+    exports: [
+        SprightChatMessageDirective
+    ]
+})
+export class SprightChatMessageModule { }